feat: Introduce single browser mode
